extract SccpSchema in its own file
