+++ wazo_confd/plugins/endpoint_sccp/resource.py
@@ -2,23 +2,13 @@
 # SPDX-License-Identifier: GPL-3.0-or-later
 
 from flask import url_for
-from marshmallow import fields
-from marshmallow.validate import Length
 
 from xivo_dao.alchemy.sccpline import SCCPLine as SCCPEndpoint
 
 from wazo_confd.auth import required_acl
-from wazo_confd.helpers.mallow import BaseSchema, Link, ListLink
 from wazo_confd.helpers.restful import ListResource, ItemResource
 
-
-class SccpSchema(BaseSchema):
-    id = fields.Integer(dump_only=True)
-    tenant_uuid = fields.String(dump_only=True)
-    options = fields.List(fields.List(fields.String(), validate=Length(equal=2)))
-    links = ListLink(Link('endpoint_sccp'))
-
-    line = fields.Nested('LineSchema', only=['id', 'links'], dump_only=True)
+from .schema import SccpSchema
 
 
 class SccpList(ListResource):
