+++ wazo_confd/plugins/endpoint_sccp/schema.py
@@ -1,0 +1,16 @@
+# Copyright 2019 The Wazo Authors  (see the AUTHORS file)
+# SPDX-License-Identifier: GPL-3.0-or-later
+
+from marshmallow import fields
+from marshmallow.validate import Length
+
+from wazo_confd.helpers.mallow import BaseSchema, Link, ListLink
+
+
+class SccpSchema(BaseSchema):
+    id = fields.Integer(dump_only=True)
+    tenant_uuid = fields.String(dump_only=True)
+    options = fields.List(fields.List(fields.String(), validate=Length(equal=2)))
+    links = ListLink(Link('endpoint_sccp'))
+
+    line = fields.Nested('LineSchema', only=['id', 'links'], dump_only=True)
